aige-backend/requirements.txt
Django>=4.2
psycopg2-binary
python-decouple
djangorestframework
django-cors-headers
djangorestframework-simplejwt
google-generativeai==0.3.2





